requirements.txt
astroid==2.9.2
black==21.12b0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
future==0.18.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pylint==2.12.2
python-whois==0.7.3
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
wrapt==1.13.3
